Event UI: Move the changeFragment method into its own class for easier sharing
