+++ app/src/main/java/ch/epfl/sweng/eventmanager/ui/event/interaction/EventAdministrationActivity.java
@@ -26,7 +26,7 @@ import ch.epfl.sweng.eventmanager.ui.event.interaction.fragments.schedule.Schedu
 import ch.epfl.sweng.eventmanager.viewmodel.ViewModelFactory;
 import dagger.android.AndroidInjection;
 
-public class EventAdministrationActivity extends AppCompatActivity
+public class EventAdministrationActivity extends MultiFragmentActivity
         implements NavigationView.OnNavigationItemSelectedListener {
 
     private static final String TAG = "EventAdministration";
@@ -117,45 +117,5 @@ public class EventAdministrationActivity extends AppCompatActivity
         }
 
         return true;
-    }
-
-    /**
-     * Change the current displayed fragment by a new one.
-     * - if the fragment is in backstack, it will pop it
-     * - if the fragment is already displayed (trying to change the fragment with the same), it will not do anything
-     *
-     * @param frag            the new fragment to display
-     * @param saveInBackstack if we want the fragment to be in backstack
-     */
-    private void changeFragment(Fragment frag, boolean saveInBackstack) {
-        String backStateName = ((Object) frag).getClass().getName();
-
-        try {
-            FragmentManager manager = getSupportFragmentManager();
-            boolean fragmentPopped = manager.popBackStackImmediate(backStateName, 0);
-
-            if (!fragmentPopped && manager.findFragmentByTag(backStateName) == null) {
-                // fragment not in back stack, create it.
-                FragmentTransaction transaction = manager.beginTransaction();
-
-                transaction.replace(R.id.content_frame, frag, backStateName);
-
-                if (saveInBackstack) {
-                    Log.d(TAG, "Change Fragment: addToBackTack " + backStateName);
-                    transaction.addToBackStack(backStateName);
-                } else {
-                    Log.d(TAG, "Change Fragment: NO addToBackTack");
-                }
-
-                transaction.commit();
-            } else {
-                // custom effect if fragment is already instanciated
-            }
-        } catch (IllegalStateException exception) {
-            Log.w(TAG,
-                    "Unable to commit fragment, could be activity as been killed in background. "
-                            + exception.toString()
-            );
-        }
     }
 }
