+++ app/src/main/java/ch/epfl/sweng/eventmanager/ui/event/interaction/EventShowcaseActivity.java
@@ -38,8 +38,8 @@ import ch.epfl.sweng.eventmanager.users.Session;
 import ch.epfl.sweng.eventmanager.viewmodel.ViewModelFactory;
 import dagger.android.AndroidInjection;
 
-public class EventShowcaseActivity extends AppCompatActivity
-        implements NavigationView.OnNavigationItemSelectedListener {
+public class EventShowcaseActivity extends MultiFragmentActivity
+        implements NavigationView.OnNavigationItemSelectedListener{
     private static final String TAG = "EventShowcaseActivity";
 
     @Inject
@@ -125,6 +125,10 @@ public class EventShowcaseActivity extends AppCompatActivity
                             return;
                         }
 
+                        // Set window title
+                        setTitle("Admin: " + ev.getName());
+
+                        // Only display the admin button is the user is cleared for ADMIN rights
                         if (Session.isLoggedIn() && Session.isClearedFor(Role.ADMIN, ev)) {
                             MenuItem adminMenuItem = navigationView.getMenu().findItem(R.id.nav_admin);
                             adminMenuItem.setVisible(true);
@@ -215,46 +219,6 @@ public class EventShowcaseActivity extends AppCompatActivity
                         navigationView.setCheckedItem(R.id.nav_tickets);
                 }
         );
-    }
-
-    /**
-     * Change the current displayed fragment by a new one.
-     * - if the fragment is in backstack, it will pop it
-     * - if the fragment is already displayed (trying to change the fragment with the same), it will not do anything
-     *
-     * @param frag            the new fragment to display
-     * @param saveInBackstack if we want the fragment to be in backstack
-     */
-    public void changeFragment(Fragment frag, boolean saveInBackstack) {
-        String backStateName = ((Object) frag).getClass().getName();
-
-        try {
-            FragmentManager manager = getSupportFragmentManager();
-            boolean fragmentPopped = manager.popBackStackImmediate(backStateName, 0);
-
-            if (!fragmentPopped && manager.findFragmentByTag(backStateName) == null) {
-                // fragment not in back stack, create it.
-                FragmentTransaction transaction = manager.beginTransaction();
-
-                transaction.replace(R.id.content_frame, frag, backStateName);
-
-                if (saveInBackstack) {
-                    Log.d(TAG, "Change Fragment: addToBackTack " + backStateName);
-                    transaction.addToBackStack(backStateName);
-                } else {
-                    Log.d(TAG, "Change Fragment: NO addToBackTack");
-                }
-
-                transaction.commit();
-            } else {
-                // TODO custom effect if fragment is already instanciated
-            }
-        } catch (IllegalStateException exception) {
-            Log.w(TAG,
-                    "Unable to commit fragment, could be activity as been killed in background. "
-                            + exception.toString()
-            );
-        }
     }
 
     /**
