+++ app/src/main/java/ch/epfl/sweng/eventmanager/ui/event/interaction/MultiFragmentActivity.java
@@ -1,0 +1,56 @@
+package ch.epfl.sweng.eventmanager.ui.event.interaction;
+
+import android.support.v4.app.Fragment;
+import android.support.v4.app.FragmentManager;
+import android.support.v4.app.FragmentTransaction;
+import android.support.v7.app.AppCompatActivity;
+import android.util.Log;
+
+import ch.epfl.sweng.eventmanager.R;
+
+import static android.support.constraint.Constraints.TAG;
+
+/**
+ * This class is used to share the changeFragment method between our various multi-fragment classes.
+ */
+public class MultiFragmentActivity extends AppCompatActivity {
+
+    /**
+     * Change the current displayed fragment by a new one.
+     * - if the fragment is in backstack, it will pop it
+     * - if the fragment is already displayed (trying to change the fragment with the same), it will not do anything
+     *  @param frag            the new fragment to display
+     * @param saveInBackstack if we want the fragment to be in backstack
+     */
+    public void changeFragment(Fragment frag, boolean saveInBackstack) {
+        String backStateName = ((Object) frag).getClass().getName();
+
+        try {
+            FragmentManager manager = getSupportFragmentManager();
+            boolean fragmentPopped = manager.popBackStackImmediate(backStateName, 0);
+
+            if (!fragmentPopped && manager.findFragmentByTag(backStateName) == null) {
+                // fragment not in back stack, create it.
+                FragmentTransaction transaction = manager.beginTransaction();
+
+                transaction.replace(R.id.content_frame, frag, backStateName);
+
+                if (saveInBackstack) {
+                    Log.d(TAG, "Change Fragment: addToBackTack " + backStateName);
+                    transaction.addToBackStack(backStateName);
+                } else {
+                    Log.d(TAG, "Change Fragment: NO addToBackTack");
+                }
+
+                transaction.commit();
+            } else {
+                // custom effect if fragment is already instanciated
+            }
+        } catch (IllegalStateException exception) {
+            Log.w(TAG,
+                    "Unable to commit fragment, could be activity as been killed in background. "
+                            + exception.toString()
+            );
+        }
+    }
+}
